updated with a new function - check_input_shapefile
check_input_shapefile allows the user to validate their shapefile prior to using forsysXR / forsysX. It checks if the fields are in the correct format (e.g. stand id as integer, etc). It will update the shapefile, either by overwriting it in the user pc or by overwrite it in the R global environment